fix(groups): address code review findings #91
- Fix star penalty leak in group-filtered leaderboard
- Move auto-association from GET provider to POST processor
- Add ON DELETE SET NULL for Session.playerGroup FK
- Update tests for creation-time auto-association
