+++ backend/src/State/SessionCreateProcessor.php
@@ -7,6 +7,8 @@ namespace App\State;
 use ApiPlatform\Doctrine\Common\State\PersistProcessor;
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\State\ProcessorInterface;
+use App\Entity\Player;
+use App\Entity\PlayerGroup;
 use App\Entity\Session;
 use Doctrine\ORM\EntityManagerInterface;
 
@@ -48,10 +50,51 @@ final readonly class SessionCreateProcessor implements ProcessorInterface
         $firstPlayer = $session->getPlayers()->first();
         if (false !== $firstPlayer) {
             $session->setCurrentDealer($firstPlayer);
-            $this->em->flush();
         }
 
+        $this->autoAssociatePlayerGroup($session, $playerIds);
+        $this->em->flush();
+
         return $session;
+    }
+
+    /**
+     * @param int[] $playerIds
+     */
+    private function autoAssociatePlayerGroup(Session $session, array $playerIds): void
+    {
+        if (empty($playerIds)) {
+            return;
+        }
+
+        $count = \count($playerIds);
+
+        /** @var list<PlayerGroup> $matchingGroups */
+        $matchingGroups = $this->em->createQuery(
+            'SELECT pg FROM App\Entity\PlayerGroup pg
+             JOIN pg.players p
+             WHERE p.id IN (:playerIds)
+             GROUP BY pg.id
+             HAVING COUNT(DISTINCT p.id) = :count'
+        )
+            ->setParameter('count', $count)
+            ->setParameter('playerIds', $playerIds)
+            ->getResult();
+
+        $matchingGroups = \array_values(\array_filter(
+            $matchingGroups,
+            static function (PlayerGroup $pg) use ($playerIds): bool {
+                $groupPlayerIds = $pg->getPlayers()->map(
+                    static fn (Player $p) => $p->getId()
+                )->getValues();
+
+                return empty(\array_diff($playerIds, $groupPlayerIds));
+            }
+        ));
+
+        if (1 === \count($matchingGroups)) {
+            $session->setPlayerGroup($matchingGroups[0]);
+        }
     }
 
     /**
